fix #806 Chess.com v3: improve text set for lessons & add ability to retry lesson via API  - adjust scrolling for texts
fix #816 Chess.com v3: rework tactics recent problems list
 - add ability to view solved tactic from recent problems list + adjust logic to prevent re-solving
 - improve data update logic for recent problems list

fix lessons categories default section selection

+++ src/com/chess/backend/entity/api/LessonProblemItem.java
@@ -439,7 +439,7 @@ public class LessonProblemItem extends BaseResponseItem<LessonProblemItem.Data> 
 		private String initial_score = "0";
 		private String last_score = "0";
 		private int current_position;
-		private int current_points;
+		private float current_points;
 		private float current_position_points;
 		/* Local addition */
 		private long lessonId;
@@ -476,7 +476,7 @@ public class LessonProblemItem extends BaseResponseItem<LessonProblemItem.Data> 
 			return current_position;
 		}
 
-		public int getCurrentPoints() {
+		public float getCurrentPoints() {
 			return current_points;
 		}
 
@@ -496,7 +496,7 @@ public class LessonProblemItem extends BaseResponseItem<LessonProblemItem.Data> 
 			this.current_position = current_position;
 		}
 
-		public void setCurrentPoints(int current_points) {
+		public void setCurrentPoints(float current_points) {
 			this.current_points = current_points;
 		}
